+++ eth_defi/erc_4626/vault_protocol/aera/__init__.py
@@ -1,0 +1,1 @@
+"""Aera vault protocol support."""